refactor(flow): group inspection internals
